Add files via upload
# Weekday Tracker

This C++ program implements a Weekday Tracker class that allows users to manipulate and track weekdays. The program includes functionalities to set a specific day, print the current day, find the next and previous days, and determine a certain day after adding a specified number of days. Built with object-oriented principles for modularity and ease of use.

## Features

- Set a specific day of the week.
- Print the current day.
- Determine the next and previous days.
- Find a certain day after adding a specified number of days.

## Usage

1. Compile the code using a C++ compiler.
2. Execute the compiled binary.
3. Follow the on-screen prompts to interact with the Weekday Tracker.

## Contributing

Contributions to enhance functionalities, improve code quality, or fix issues are welcome! Please fork the repository, make your changes, and submit a pull request.

## License

This project is licensed under the [irfan shakil]